+++ source/c_src/Interface/include/HSpectrumHeaderStruct.h
@@ -20,17 +20,17 @@
 
 struct HSpectrumHeaderStruct
 {
-    size_t fHeaderSize; //size of the header struct in bytes
+    uint64_t fHeaderSize; //size of the header struct in bytes
     char fVersionFlag[HFLAG_WIDTH]; //version, format code (also indicates what type (e.g. double, float) the spectrum data is stored in)
     char fSidebandFlag[HFLAG_WIDTH]; //flag indicating the sideband
     char fPolarizationFlag[HFLAG_WIDTH]; //flag indicating the polarization recorded
     uint64_t fStartTime; //acquisition start time in seconds since epoch
     uint64_t fSampleRate; //may need to accomodate doubles?
     uint64_t fLeadingSampleIndex; //sample index since start of the acquisition
-    size_t fSampleLength; //total number of samples used to compute the spectrum
-    size_t fNAverages; //if spectral averaging was used, how many averaging periods were used
-    size_t fSpectrumLength; //number of spectral points
-    size_t fSpectrumDataTypeSize; //spectrum data point size (e.g. size of data type storing the spectrum; sizeof(float), etc)
+    uint64_t fSampleLength; //total number of samples used to compute the spectrum
+    uint64_t fNAverages; //if spectral averaging was used, how many averaging periods were used
+    uint64_t fSpectrumLength; //number of spectral points
+    uint64_t fSpectrumDataTypeSize; //spectrum data point size (e.g. size of data type storing the spectrum; sizeof(float), etc)
     char fExperimentName[HNAME_WIDTH]; //experiment name
     char fSourceName[HNAME_WIDTH]; //source name
     char fScanName[HNAME_WIDTH]; //scan name
